Extract presenter from test
